optimizado el coio ahora teemos servicios donde ponemos todo y los contollers estan limpios como la api
