+++ cine_web_app/back_end/Models/Cine.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 
 namespace cine_web_app.back_end.Models
 {
@@ -10,4 +11,4 @@ namespace cine_web_app.back_end.Models
         // Agregar la propiedad Peliculas para almacenar la lista de películas en cada cine
         public List<Pelicula> Peliculas { get; set; }
     }
-}   
+}
